requirements.txt
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
contourpy==1.3.3
curl_cffi==0.15.0
cycler==0.12.1
distro==1.9.0
fonttools==4.62.1
frozendict==2.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
kiwisolver==1.5.0
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
multitasking==0.0.12
numpy==2.4.4
openai==2.30.0
packaging==26.0
pandas==3.0.2
peewee==4.0.4
pillow==12.2.0
platformdirs==4.9.4
protobuf==7.34.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
rich==14.3.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
ta==0.11.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yfinance==1.2.1
